Resolved merge conflict

+++ init.sh
@@ -4,4 +4,6 @@ set -ex
 
 mkdir -p secrets ctf_data/db_backups ctf_data/files ctf_data/restore oauth db_data
 
+if [[ ! -e secrets/db_password ]]; then
 cat /dev/urandom | tr -dc 'a-zA-Z0-9\!\@\#\$\%\^\&\*\(\)\_' | fold -w 32 | head -n 1 > secrets/db_password
+fi
